db improvements

+++ src/main/resources/db/migration/V20251022_020550__create_table_book_aud.sql
@@ -1,6 +1,6 @@
 CREATE TABLE ${audit_schema_name}.book_aud (
-  id            INT           NOT NULL,
-  rev           INT           NOT NULL,
+  id            BIGINT        NOT NULL,
+  rev           BIGINT        NOT NULL,
   revtype       SMALLINT      NOT NULL,
   rev_timestamp TIMESTAMP     NOT NULL DEFAULT CURRENT_TIMESTAMP,
   title         VARCHAR(255)      NULL,
